chapter_002/src/test/java/ru/job4j/paint/SquareTest.java
package ru.job4j.paint;

import org.junit.Test;
import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;

/**
 * Test Square.
 * @author Denis Seleznev (d.selezneww@mail.ru)
 * @version $Id$
 * @since 0.1
 */
public class SquareTest {

    /**
     * Test draw.
     */
    @Test
    public void whenDrawSquare() {
        Square square = new Square();
        assertThat(
                square.draw(),
                is(
                        new StringBuilder()
                                .append("†††††††††")
                                .append("†     †")
                                .append("†     †")
                                .append("†††††††††")
                                .toString()
                )
        );
    }
}

